requirements.txt
wheel
gevent>=1.1.0
bitcoin
ipaddress
coverage
tox
coincurve>=5.0.1
pycryptodome>=3.3.1
rlp>=0.5.1,<0.6.0
miniupnpc
repoze.lru
cryptography>=1.8.0
